Hose entire project
So I thought the problem was a fundametnal problem with the HttpClient engines
in ktor. Therefore I hosed the project to rebuild in C#, since I am much more
familiar with the .NET/.NET Core environment due to my work experience.

The problem to which I refer is where half of the data is not returned (name,
company, blog, etc.) And I have spent most of the past time since my last
commit trying to fix it. This is my resignation to try to get something that
completely works out and functional.

Spoiler alert: it didn't work. The object still doesn't return fully, even in
.NET Core's HttpClient. I now think it has something to do with https and the
fact that I do not have https set up on my home apis, which would explain why
it works through my web browser and node.js Axios calls.